Updated and completed documentation (including comments)
