+++ code/hand_label_assistant/label_dialog.py
@@ -1,440 +1,494 @@
-# -*- coding: utf-8 -*-
+from PyQt5.QtCore import *
+from PyQt5.QtGui import *
+from PyQt5.QtWidgets import *
 
-# Form implementation generated from reading ui file 'label_dialog.ui'
-#
-# Created by: PyQt5 UI code generator 5.13.0
-#
-# WARNING! All changes made in this file will be lost!
+import os
+import sys
+
+import traceback
+
+from label_dialog_ui import *
+import numpy as np
+
+import numpy as np
+import pandas as pd
+import imageio
+
+import re
+from feature_extraction.feature_extraction import *
+from feature_extraction.color_plot import color_plot
+
+from pandas_model import PandasModel
 
 
-from PyQt5 import QtCore, QtGui, QtWidgets
+class LabelingDialog(QWidget):
+    def __init__(self, widget_handled, ui):
+        """ Initializes the labeling Dialog
+        Args:
+            widget_handeled: Events for this widget are handeled by LabelingDialog to access arrow keys
+            ui: User interface generated by pyuic5 (view)
+        """
+        self.thread = LabelingDialog.Features(self)
+        self.extract_features = False
+        self.ui = ui
+        self.make_conncections()
+        self.widget_handled = widget_handled #Widget for event Handling. All events are checked and if not arrow keys passed on. See eventFilter below
+        QWidget.__init__(self, widget_handled)
+
+        self.outpath = None#Thr path of the output csv
+        self.labels = None#After loading a dictionary that contains: key=*index of asparagus* to value=*list of properties*
+
+        self.idx_image = 0#Relates to the index of the current asparagus piece i.e. the index of the corresponding three files
+        self.images = {}
+        self.ui.color.set_gray_background()
 
 
-class Ui_LabelDialog(object):
-    def setupUi(self, LabelDialog):
-        LabelDialog.setObjectName("LabelDialog")
-        LabelDialog.resize(1425, 980)
-        self.centralwidget = QtWidgets.QWidget(LabelDialog)
-        self.centralwidget.setObjectName("centralwidget")
-        self.horizontalLayout_3 = QtWidgets.QHBoxLayout(self.centralwidget)
-        self.horizontalLayout_3.setObjectName("horizontalLayout_3")
-        self.line = QtWidgets.QFrame(self.centralwidget)
-        self.line.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line.setObjectName("line")
-        self.horizontalLayout_3.addWidget(self.line)
-        self.frame = QtWidgets.QFrame(self.centralwidget)
-        self.frame.setFrameShape(QtWidgets.QFrame.StyledPanel)
-        self.frame.setFrameShadow(QtWidgets.QFrame.Raised)
-        self.frame.setObjectName("frame")
-        self.verticalLayout = QtWidgets.QVBoxLayout(self.frame)
-        self.verticalLayout.setObjectName("verticalLayout")
-        self.label = ImageDisplay(self.frame)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Preferred)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.label.sizePolicy().hasHeightForWidth())
-        self.label.setSizePolicy(sizePolicy)
-        self.label.setText("")
-        self.label.setObjectName("label")
-        self.verticalLayout.addWidget(self.label)
-        self.line_2 = QtWidgets.QFrame(self.frame)
-        self.line_2.setFrameShape(QtWidgets.QFrame.HLine)
-        self.line_2.setFrameShadow(QtWidgets.QFrame.Sunken)
-        self.line_2.setObjectName("line_2")
-        self.verticalLayout.addWidget(self.line_2)
-        self.widget = QtWidgets.QWidget(self.frame)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Maximum)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.widget.sizePolicy().hasHeightForWidth())
-        self.widget.setSizePolicy(sizePolicy)
-        self.widget.setObjectName("widget")
-        self.verticalLayout_2 = QtWidgets.QVBoxLayout(self.widget)
-        self.verticalLayout_2.setObjectName("verticalLayout_2")
-        self.question = QtWidgets.QLineEdit(self.widget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.question.sizePolicy().hasHeightForWidth())
-        self.question.setSizePolicy(sizePolicy)
-        self.question.setObjectName("question")
-        self.verticalLayout_2.addWidget(self.question)
-        self.widget_2 = QtWidgets.QWidget(self.widget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Preferred, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.widget_2.sizePolicy().hasHeightForWidth())
-        self.widget_2.setSizePolicy(sizePolicy)
-        self.widget_2.setObjectName("widget_2")
-        self.horizontalLayout = QtWidgets.QHBoxLayout(self.widget_2)
-        self.horizontalLayout.setObjectName("horizontalLayout")
-        self.asparagus_name = QtWidgets.QLineEdit(self.widget_2)
-        self.asparagus_name.setMinimumSize(QtCore.QSize(150, 0))
-        self.asparagus_name.setMaximumSize(QtCore.QSize(150, 16777215))
-        self.asparagus_name.setObjectName("asparagus_name")
-        self.horizontalLayout.addWidget(self.asparagus_name)
-        self.asparagus_no = QtWidgets.QSpinBox(self.widget_2)
-        self.asparagus_no.setMinimumSize(QtCore.QSize(80, 0))
-        self.asparagus_no.setMaximumSize(QtCore.QSize(100, 16777215))
-        self.asparagus_no.setMaximum(999999999)
-        self.asparagus_no.setObjectName("asparagus_no")
-        self.horizontalLayout.addWidget(self.asparagus_no)
-        spacerItem = QtWidgets.QSpacerItem(40, 20, QtWidgets.QSizePolicy.Expanding, QtWidgets.QSizePolicy.Minimum)
-        self.horizontalLayout.addItem(spacerItem)
-        self.resetCurrentAsparagus = QtWidgets.QPushButton(self.widget_2)
-        self.resetCurrentAsparagus.setObjectName("resetCurrentAsparagus")
-        self.horizontalLayout.addWidget(self.resetCurrentAsparagus)
-        self.previous_question = QtWidgets.QPushButton(self.widget_2)
-        self.previous_question.setMaximumSize(QtCore.QSize(35, 16777215))
-        self.previous_question.setObjectName("previous_question")
-        self.horizontalLayout.addWidget(self.previous_question)
-        self.next_question = QtWidgets.QPushButton(self.widget_2)
-        self.next_question.setMaximumSize(QtCore.QSize(35, 16777215))
-        self.next_question.setObjectName("next_question")
-        self.horizontalLayout.addWidget(self.next_question)
-        self.yes = QtWidgets.QPushButton(self.widget_2)
-        self.yes.setObjectName("yes")
-        self.horizontalLayout.addWidget(self.yes)
-        self.no = QtWidgets.QPushButton(self.widget_2)
-        self.no.setObjectName("no")
-        self.horizontalLayout.addWidget(self.no)
-        self.verticalLayout_2.addWidget(self.widget_2)
-        self.verticalLayout.addWidget(self.widget)
-        self.horizontalLayout_3.addWidget(self.frame)
-        self.scrollArea = QtWidgets.QScrollArea(self.centralwidget)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Minimum)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.scrollArea.sizePolicy().hasHeightForWidth())
-        self.scrollArea.setSizePolicy(sizePolicy)
-        self.scrollArea.setMinimumSize(QtCore.QSize(200, 0))
-        self.scrollArea.setMaximumSize(QtCore.QSize(300, 16777215))
-        self.scrollArea.setWidgetResizable(True)
-        self.scrollArea.setObjectName("scrollArea")
-        self.scrollAreaWidgetContents = QtWidgets.QWidget()
-        self.scrollAreaWidgetContents.setGeometry(QtCore.QRect(0, 0, 298, 960))
-        self.scrollAreaWidgetContents.setObjectName("scrollAreaWidgetContents")
-        self.verticalLayout_10 = QtWidgets.QVBoxLayout(self.scrollAreaWidgetContents)
-        self.verticalLayout_10.setObjectName("verticalLayout_10")
-        self.groupBoxViolet = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBoxViolet.setObjectName("groupBoxViolet")
-        self.verticalLayout_5 = QtWidgets.QVBoxLayout(self.groupBoxViolet)
-        self.verticalLayout_5.setObjectName("verticalLayout_5")
-        self.color = ImageDisplay(self.groupBoxViolet)
-        self.color.setMinimumSize(QtCore.QSize(250, 150))
-        self.color.setText("")
-        self.color.setObjectName("color")
-        self.verticalLayout_5.addWidget(self.color)
-        self.horizontalLayout_5 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_5.setObjectName("horizontalLayout_5")
-        self.predictionViolet_2 = QtWidgets.QLineEdit(self.groupBoxViolet)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionViolet_2.sizePolicy().hasHeightForWidth())
-        self.predictionViolet_2.setSizePolicy(sizePolicy)
-        self.predictionViolet_2.setMinimumSize(QtCore.QSize(10, 0))
-        self.predictionViolet_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionViolet_2.setObjectName("predictionViolet_2")
-        self.horizontalLayout_5.addWidget(self.predictionViolet_2)
-        self.predictionViolet = QtWidgets.QLineEdit(self.groupBoxViolet)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionViolet.sizePolicy().hasHeightForWidth())
-        self.predictionViolet.setSizePolicy(sizePolicy)
-        self.predictionViolet.setMinimumSize(QtCore.QSize(10, 0))
-        self.predictionViolet.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionViolet.setObjectName("predictionViolet")
-        self.horizontalLayout_5.addWidget(self.predictionViolet)
-        self.predictionViolet_3 = QtWidgets.QLineEdit(self.groupBoxViolet)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionViolet_3.sizePolicy().hasHeightForWidth())
-        self.predictionViolet_3.setSizePolicy(sizePolicy)
-        self.predictionViolet_3.setMinimumSize(QtCore.QSize(10, 0))
-        self.predictionViolet_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionViolet_3.setObjectName("predictionViolet_3")
-        self.horizontalLayout_5.addWidget(self.predictionViolet_3)
-        self.verticalLayout_5.addLayout(self.horizontalLayout_5)
-        self.overallPredictionViolet = QtWidgets.QLineEdit(self.groupBoxViolet)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictionViolet.sizePolicy().hasHeightForWidth())
-        self.overallPredictionViolet.setSizePolicy(sizePolicy)
-        self.overallPredictionViolet.setObjectName("overallPredictionViolet")
-        self.verticalLayout_5.addWidget(self.overallPredictionViolet)
-        self.usePredictionViolet = QtWidgets.QCheckBox(self.groupBoxViolet)
-        self.usePredictionViolet.setObjectName("usePredictionViolet")
-        self.verticalLayout_5.addWidget(self.usePredictionViolet)
-        self.overallPredictionViolet.raise_()
-        self.usePredictionViolet.raise_()
-        self.color.raise_()
-        self.verticalLayout_10.addWidget(self.groupBoxViolet)
-        self.groupBoxBlooming = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBoxBlooming.setObjectName("groupBoxBlooming")
-        self.verticalLayout_9 = QtWidgets.QVBoxLayout(self.groupBoxBlooming)
-        self.verticalLayout_9.setObjectName("verticalLayout_9")
-        self.horizontalLayout_11 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_11.setObjectName("horizontalLayout_11")
-        self.predictionBlooming_2 = QtWidgets.QLineEdit(self.groupBoxBlooming)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBlooming_2.sizePolicy().hasHeightForWidth())
-        self.predictionBlooming_2.setSizePolicy(sizePolicy)
-        self.predictionBlooming_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBlooming_2.setObjectName("predictionBlooming_2")
-        self.horizontalLayout_11.addWidget(self.predictionBlooming_2)
-        self.predictionBlooming = QtWidgets.QLineEdit(self.groupBoxBlooming)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBlooming.sizePolicy().hasHeightForWidth())
-        self.predictionBlooming.setSizePolicy(sizePolicy)
-        self.predictionBlooming.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBlooming.setObjectName("predictionBlooming")
-        self.horizontalLayout_11.addWidget(self.predictionBlooming)
-        self.predictionBlooming_3 = QtWidgets.QLineEdit(self.groupBoxBlooming)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBlooming_3.sizePolicy().hasHeightForWidth())
-        self.predictionBlooming_3.setSizePolicy(sizePolicy)
-        self.predictionBlooming_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBlooming_3.setObjectName("predictionBlooming_3")
-        self.horizontalLayout_11.addWidget(self.predictionBlooming_3)
-        self.verticalLayout_9.addLayout(self.horizontalLayout_11)
-        self.overallPredictionBlooming = QtWidgets.QLineEdit(self.groupBoxBlooming)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictionBlooming.sizePolicy().hasHeightForWidth())
-        self.overallPredictionBlooming.setSizePolicy(sizePolicy)
-        self.overallPredictionBlooming.setObjectName("overallPredictionBlooming")
-        self.verticalLayout_9.addWidget(self.overallPredictionBlooming)
-        self.usePredictionBlooming = QtWidgets.QCheckBox(self.groupBoxBlooming)
-        self.usePredictionBlooming.setObjectName("usePredictionBlooming")
-        self.verticalLayout_9.addWidget(self.usePredictionBlooming)
-        self.verticalLayout_10.addWidget(self.groupBoxBlooming)
-        self.groupBoxLength = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBoxLength.setObjectName("groupBoxLength")
-        self.verticalLayout_7 = QtWidgets.QVBoxLayout(self.groupBoxLength)
-        self.verticalLayout_7.setObjectName("verticalLayout_7")
-        self.horizontalLayout_9 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_9.setObjectName("horizontalLayout_9")
-        self.predictionLength = QtWidgets.QLineEdit(self.groupBoxLength)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionLength.sizePolicy().hasHeightForWidth())
-        self.predictionLength.setSizePolicy(sizePolicy)
-        self.predictionLength.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionLength.setObjectName("predictionLength")
-        self.horizontalLayout_9.addWidget(self.predictionLength)
-        self.predictionLength_2 = QtWidgets.QLineEdit(self.groupBoxLength)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionLength_2.sizePolicy().hasHeightForWidth())
-        self.predictionLength_2.setSizePolicy(sizePolicy)
-        self.predictionLength_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionLength_2.setObjectName("predictionLength_2")
-        self.horizontalLayout_9.addWidget(self.predictionLength_2)
-        self.predictionLength_3 = QtWidgets.QLineEdit(self.groupBoxLength)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionLength_3.sizePolicy().hasHeightForWidth())
-        self.predictionLength_3.setSizePolicy(sizePolicy)
-        self.predictionLength_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionLength_3.setObjectName("predictionLength_3")
-        self.horizontalLayout_9.addWidget(self.predictionLength_3)
-        self.verticalLayout_7.addLayout(self.horizontalLayout_9)
-        self.overallPredictionLength = QtWidgets.QLineEdit(self.groupBoxLength)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictionLength.sizePolicy().hasHeightForWidth())
-        self.overallPredictionLength.setSizePolicy(sizePolicy)
-        self.overallPredictionLength.setObjectName("overallPredictionLength")
-        self.verticalLayout_7.addWidget(self.overallPredictionLength)
-        self.usePredictionLength = QtWidgets.QCheckBox(self.groupBoxLength)
-        self.usePredictionLength.setObjectName("usePredictionLength")
-        self.verticalLayout_7.addWidget(self.usePredictionLength)
-        self.verticalLayout_10.addWidget(self.groupBoxLength)
-        self.groupBoxRust = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBoxRust.setObjectName("groupBoxRust")
-        self.verticalLayout_4 = QtWidgets.QVBoxLayout(self.groupBoxRust)
-        self.verticalLayout_4.setObjectName("verticalLayout_4")
-        self.horizontalLayout_7 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_7.setObjectName("horizontalLayout_7")
-        self.predictionRust_2 = QtWidgets.QLineEdit(self.groupBoxRust)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionRust_2.sizePolicy().hasHeightForWidth())
-        self.predictionRust_2.setSizePolicy(sizePolicy)
-        self.predictionRust_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionRust_2.setObjectName("predictionRust_2")
-        self.horizontalLayout_7.addWidget(self.predictionRust_2)
-        self.predictionRust = QtWidgets.QLineEdit(self.groupBoxRust)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionRust.sizePolicy().hasHeightForWidth())
-        self.predictionRust.setSizePolicy(sizePolicy)
-        self.predictionRust.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionRust.setObjectName("predictionRust")
-        self.horizontalLayout_7.addWidget(self.predictionRust)
-        self.predictionRust_3 = QtWidgets.QLineEdit(self.groupBoxRust)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionRust_3.sizePolicy().hasHeightForWidth())
-        self.predictionRust_3.setSizePolicy(sizePolicy)
-        self.predictionRust_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionRust_3.setObjectName("predictionRust_3")
-        self.horizontalLayout_7.addWidget(self.predictionRust_3)
-        self.verticalLayout_4.addLayout(self.horizontalLayout_7)
-        self.overallPredictedValueRust = QtWidgets.QLineEdit(self.groupBoxRust)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictedValueRust.sizePolicy().hasHeightForWidth())
-        self.overallPredictedValueRust.setSizePolicy(sizePolicy)
-        self.overallPredictedValueRust.setObjectName("overallPredictedValueRust")
-        self.verticalLayout_4.addWidget(self.overallPredictedValueRust)
-        self.usePredictionRust = QtWidgets.QCheckBox(self.groupBoxRust)
-        self.usePredictionRust.setObjectName("usePredictionRust")
-        self.verticalLayout_4.addWidget(self.usePredictionRust)
-        self.verticalLayout_10.addWidget(self.groupBoxRust)
-        self.groupBox = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBox.setObjectName("groupBox")
-        self.verticalLayout_6 = QtWidgets.QVBoxLayout(self.groupBox)
-        self.verticalLayout_6.setObjectName("verticalLayout_6")
-        self.horizontalLayout_8 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_8.setObjectName("horizontalLayout_8")
-        self.predictionWidth = QtWidgets.QLineEdit(self.groupBox)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionWidth.sizePolicy().hasHeightForWidth())
-        self.predictionWidth.setSizePolicy(sizePolicy)
-        self.predictionWidth.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionWidth.setObjectName("predictionWidth")
-        self.horizontalLayout_8.addWidget(self.predictionWidth)
-        self.predictionWidth_2 = QtWidgets.QLineEdit(self.groupBox)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionWidth_2.sizePolicy().hasHeightForWidth())
-        self.predictionWidth_2.setSizePolicy(sizePolicy)
-        self.predictionWidth_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionWidth_2.setObjectName("predictionWidth_2")
-        self.horizontalLayout_8.addWidget(self.predictionWidth_2)
-        self.predictionWidth_3 = QtWidgets.QLineEdit(self.groupBox)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionWidth_3.sizePolicy().hasHeightForWidth())
-        self.predictionWidth_3.setSizePolicy(sizePolicy)
-        self.predictionWidth_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionWidth_3.setObjectName("predictionWidth_3")
-        self.horizontalLayout_8.addWidget(self.predictionWidth_3)
-        self.verticalLayout_6.addLayout(self.horizontalLayout_8)
-        self.overallPredictionWidth = QtWidgets.QLineEdit(self.groupBox)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictionWidth.sizePolicy().hasHeightForWidth())
-        self.overallPredictionWidth.setSizePolicy(sizePolicy)
-        self.overallPredictionWidth.setObjectName("overallPredictionWidth")
-        self.verticalLayout_6.addWidget(self.overallPredictionWidth)
-        self.usePredictionWidth = QtWidgets.QCheckBox(self.groupBox)
-        self.usePredictionWidth.setObjectName("usePredictionWidth")
-        self.verticalLayout_6.addWidget(self.usePredictionWidth)
-        self.verticalLayout_10.addWidget(self.groupBox)
-        self.groupBoxBended = QtWidgets.QGroupBox(self.scrollAreaWidgetContents)
-        self.groupBoxBended.setObjectName("groupBoxBended")
-        self.verticalLayout_8 = QtWidgets.QVBoxLayout(self.groupBoxBended)
-        self.verticalLayout_8.setObjectName("verticalLayout_8")
-        self.horizontalLayout_10 = QtWidgets.QHBoxLayout()
-        self.horizontalLayout_10.setObjectName("horizontalLayout_10")
-        self.predictionBended_3 = QtWidgets.QLineEdit(self.groupBoxBended)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBended_3.sizePolicy().hasHeightForWidth())
-        self.predictionBended_3.setSizePolicy(sizePolicy)
-        self.predictionBended_3.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBended_3.setObjectName("predictionBended_3")
-        self.horizontalLayout_10.addWidget(self.predictionBended_3)
-        self.predictionBended_2 = QtWidgets.QLineEdit(self.groupBoxBended)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBended_2.sizePolicy().hasHeightForWidth())
-        self.predictionBended_2.setSizePolicy(sizePolicy)
-        self.predictionBended_2.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBended_2.setObjectName("predictionBended_2")
-        self.horizontalLayout_10.addWidget(self.predictionBended_2)
-        self.predictionBended = QtWidgets.QLineEdit(self.groupBoxBended)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.predictionBended.sizePolicy().hasHeightForWidth())
-        self.predictionBended.setSizePolicy(sizePolicy)
-        self.predictionBended.setMaximumSize(QtCore.QSize(80, 16777215))
-        self.predictionBended.setObjectName("predictionBended")
-        self.horizontalLayout_10.addWidget(self.predictionBended)
-        self.verticalLayout_8.addLayout(self.horizontalLayout_10)
-        self.overallPredictionBended = QtWidgets.QLineEdit(self.groupBoxBended)
-        sizePolicy = QtWidgets.QSizePolicy(QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Fixed)
-        sizePolicy.setHorizontalStretch(0)
-        sizePolicy.setVerticalStretch(0)
-        sizePolicy.setHeightForWidth(self.overallPredictionBended.sizePolicy().hasHeightForWidth())
-        self.overallPredictionBended.setSizePolicy(sizePolicy)
-        self.overallPredictionBended.setObjectName("overallPredictionBended")
-        self.verticalLayout_8.addWidget(self.overallPredictionBended)
-        self.usePredictedValueBended = QtWidgets.QCheckBox(self.groupBoxBended)
-        self.usePredictedValueBended.setObjectName("usePredictedValueBended")
-        self.verticalLayout_8.addWidget(self.usePredictedValueBended)
-        self.verticalLayout_10.addWidget(self.groupBoxBended)
-        self.groupBoxViolet.raise_()
-        self.groupBoxRust.raise_()
-        self.groupBox.raise_()
-        self.groupBoxLength.raise_()
-        self.groupBoxBended.raise_()
-        self.groupBoxBlooming.raise_()
-        self.scrollArea.setWidget(self.scrollAreaWidgetContents)
-        self.horizontalLayout_3.addWidget(self.scrollArea)
-        LabelDialog.setCentralWidget(self.centralwidget)
+        headers_main_variables = ["is_bruch","is_hollow","has_blume","has_rost_head","has_rost_body","is_bended","is_violet","very_thick","thick","medium_thick","thin","very_thin","unclassified"]
+        headers_set_via_feature_extraction = ["auto_violet","auto_blooming","auto_length","auto_rust_head","auto_rust_body","auto_width","auto_bended"]
+        headers_additional_extracted_features = []
 
-        self.retranslateUi(LabelDialog)
-        QtCore.QMetaObject.connectSlotsByName(LabelDialog)
+        self.outfile_headers = []
+        self.outfile_headers.extend(headers_main_variables)
+        self.outfile_headers.extend(headers_set_via_feature_extraction)
+        self.outfile_headers.extend(headers_additional_extracted_features)
 
-    def retranslateUi(self, LabelDialog):
-        _translate = QtCore.QCoreApplication.translate
-        LabelDialog.setWindowTitle(_translate("LabelDialog", "Hand Label Assistant"))
-        self.resetCurrentAsparagus.setText(_translate("LabelDialog", "Reset current asparagus"))
-        self.previous_question.setText(_translate("LabelDialog", "<<"))
-        self.next_question.setText(_translate("LabelDialog", ">>"))
-        self.yes.setText(_translate("LabelDialog", "Yes"))
-        self.no.setText(_translate("LabelDialog", "No"))
-        self.groupBoxViolet.setTitle(_translate("LabelDialog", "Violet"))
-        self.usePredictionViolet.setText(_translate("LabelDialog", "Use predicted value"))
-        self.groupBoxBlooming.setTitle(_translate("LabelDialog", "Blooming"))
-        self.usePredictionBlooming.setText(_translate("LabelDialog", "Use predicted value"))
-        self.groupBoxLength.setTitle(_translate("LabelDialog", "Length"))
-        self.usePredictionLength.setText(_translate("LabelDialog", "Use predicted value"))
-        self.groupBoxRust.setTitle(_translate("LabelDialog", "Rust"))
-        self.usePredictionRust.setText(_translate("LabelDialog", "Use predicted value"))
-        self.groupBox.setTitle(_translate("LabelDialog", "Width"))
-        self.usePredictionWidth.setText(_translate("LabelDialog", "Use predicted value"))
-        self.groupBoxBended.setTitle(_translate("LabelDialog", "Bended"))
-        self.usePredictedValueBended.setText(_translate("LabelDialog", "Use predicted value"))
-from imagedisplay import ImageDisplay
+        self.questions = headers_main_variables[:-1]#Questions must be the first rows in out
+
+
+        self.feature_to_questions = { "width":["very_thick","thick","medium_thick","thin","very_thin"],
+                                 "blooming":["has_blume"],
+                                 "length":["is_bruch"],
+                                 "rust":["has_rost_head","has_rost_body"],
+                                 "violet":["is_violet"],
+                                 "bended":["is_bended"]
+                                }
+        self.automatic = []
+
+        self.idx_question = 0
+        self.ui.question.setText(self.questions[self.idx_question])
+        self.update_info()
+
+    def not_classifiable(self):
+        """ Sets a value in the outputfile that indicates the asparagus piece was not classifiable"""
+        try:
+            self.set_value_for_label(1, "unclassified")
+            self.next_image()
+        except e as Exception:
+            print(e)
+
+    def use_feature_extraction_for(self, feature):
+        """ Specifies for which features the user wants to rely on automaticlly extracted features.
+            Questions for the feature being extracted automatically are not posed to the user for manual labeling.
+            Args:
+                feature: The feature (see: self.feature_to_questions.keys()) that should not be estimated by the user manually.
+        """
+        remove = self.feature_to_questions[feature]
+        for x in remove:
+            self.questions.remove(x)
+
+    def use_question_for(self, feature):
+        """ Specifies features the user wants questions to be asked for. Adds the respective questions to add to the list of questions asked for labeling (self.questions) """
+        self.questions.extend(self.feature_to_questions[feature])
+
+    def make_conncections(self):
+        """ Establish connections between UI elements and functionalities"""
+        self.ui.asparagus_number.valueChanged.connect(lambda x: self.set_index(int(x)))
+        self.ui.previous_question.clicked.connect(self.previous_question)
+        self.ui.next_question.clicked.connect(self.next_question)
+        self.ui.yes.clicked.connect(self.answer_yes)
+        self.ui.no.clicked.connect(self.answer_no)
+        self.ui.notClassifiable.clicked.connect(self.not_classifiable)
+
+        self.ui.usePredictionLength.stateChanged.connect(lambda x: self.use_feature_extraction_for("length") if x else self.use_question_for("length"))
+        self.ui.usePredictionBlooming.stateChanged.connect(lambda x: self.use_feature_extraction_for("blooming") if x else self.use_question_for("blooming"))
+        self.ui.usePredictionRust.stateChanged.connect(lambda x: self.use_feature_extraction_for("rust") if x else self.use_question_for("rust"))
+        self.ui.usePredictionWidth.stateChanged.connect(lambda x: self.use_feature_extraction_for("width") if x else self.use_question_for("width"))
+        self.ui.usePredictionBended.stateChanged.connect(lambda x: self.use_feature_extraction_for("bended") if x else self.use_question_for("bended"))
+        self.ui.usePredictionViolet.stateChanged.connect(lambda x: self.use_feature_extraction_for("violet") if x else self.use_question_for("violet"))
+
+        self.ui.extractFeatures.toggled.connect(self.toggle_feature_extraction)
+
+        self.thread.color_plot.connect(self.ui.color.update)
+
+        # The following connections are made to link the results of the feature extraction thread to ui elements or the methods of self thats save them to the output csv file.
+        self.thread.predictionWidth.connect(self.ui.predictionWidth.setText)
+        self.thread.predictionBended.connect(self.ui.predictionBended.setText)
+        self.thread.predictionLength.connect(self.ui.predictionLength.setText)
+        self.thread.predictionViolet.connect(self.ui.predictionViolet.setText)
+        self.thread.predictionRust.connect(self.ui.predictionRust.setText)
+        self.thread.predictionBlooming.connect(self.ui.predictionBlooming.setText)
+
+        self.thread.predictionWidth_2.connect(self.ui.predictionWidth_2.setText)
+        self.thread.predictionBended_2.connect(self.ui.predictionBended_2.setText)
+        self.thread.predictionLength_2.connect(self.ui.predictionLength_2.setText)
+        self.thread.predictionViolet_2.connect(self.ui.predictionViolet_2.setText)
+        self.thread.predictionRust_2.connect(self.ui.predictionRust_2.setText)
+        self.thread.predictionBlooming_2.connect(self.ui.predictionBlooming_2.setText)
+
+        self.thread.predictionWidth_3.connect(self.ui.predictionWidth_3.setText)
+        self.thread.predictionBended_3.connect(self.ui.predictionBended_3.setText)
+        self.thread.predictionLength_3.connect(self.ui.predictionLength_3.setText)
+        self.thread.predictionViolet_3.connect(self.ui.predictionViolet_3.setText)
+        self.thread.predictionRust_3.connect(self.ui.predictionRust_3.setText)
+        self.thread.predictionBlooming_3.connect(self.ui.predictionBlooming_3.setText)
+
+        self.thread.overallPredictionWidth.connect(self.ui.overallPredictionWidth.setText)
+        self.thread.overallPredictionWidth.connect(lambda x: self.set_value_for_label(x, "auto_width",self.idx_image))
+        self.thread.overallPredictionBended.connect(self.ui.overallPredictionBended.setText)
+        self.thread.overallPredictionBended.connect(lambda x: self.set_value_for_label(x, "auto_bended",self.idx_image))
+        self.thread.overallPredictionLength.connect(self.ui.overallPredictionLength.setText)
+        self.thread.overallPredictionLength.connect(lambda x: self.set_value_for_label(x, "auto_length",self.idx_image))
+        self.thread.overallPredictionViolet.connect(self.ui.overallPredictionViolet.setText)
+        self.thread.overallPredictionViolet.connect(lambda x: self.set_value_for_label(x, "auto_violet",self.idx_image))
+        self.thread.overallPredictedValueRust.connect(self.ui.overallPredictedValueRust.setText)
+        self.thread.overallPredictionBlooming.connect(self.ui.overallPredictionBlooming.setText)
+
+    def toggle_feature_extraction(self):
+        """ Toggles the automatic extraction of features """
+        self.extract_features = not self.extract_features
+        self.set_index(self.idx_image)
+
+    def set_output_file(self, path):
+        """ Sets output file
+            Args:
+                path
+        """
+        self.outpath = path
+        self.load_outfile()
+
+    def load_outfile(self):
+        """ Loads outputfile if it exists and saves contents to self.dict.
+            Sets self.labels to an empty dict otherwise."""
+        try:
+            recovered = pd.read_csv(self.outpath, index_col=0, sep =";").to_dict(orient="index")
+            for key, value in recovered.items():
+                recovered[key] = list(recovered[key].values())
+            self.labels = recovered
+        except FileNotFoundError:
+            self.labels = {}
+        except e as Exception:
+            print(e)
+
+    def answer_yes(self):
+        """ Writes answer (yes) to current question to file """
+        self.log_answer(1)
+        self.ui.yes.setFocus()
+        self.next_question()
+
+    def answer_no(self):
+        """ Writes answer (no) to current question to file """
+        self.log_answer(0)
+        self.ui.no.setFocus()
+        self.next_question()
+
+    def set_value_for_label(self, value, label, idx = None):
+        """ Sets value for a given label. Exception passed on if self.labels is not set yet.
+        Args:
+            value: The value the feature should be set to
+            label: The label tha value of which shall be updated
+            idx: The index of the aspargus piece. If no value is specified the index of the asparagus piece currently being displayed is used.
+        """
+        if not type(self.labels) == type({}):
+            QMessageBox.about(self, "Attention", "Load output file first.")
+            return
+        #print("set value")
+
+        if idx == None:
+            idx = self.idx_image
+
+        try:#Assure we have a line of data for the current index
+            self.labels[idx]
+        except KeyError:
+            self.labels[idx] = [None for x in range(len(self.outfile_headers))]
+
+        try:
+            self.labels[idx][self.outfile_headers.index(label)] = value
+        except e as Exception:
+            print("Couldn't set label")
+            print(e)
+
+    def log_answer(self, answer):
+        """ Logs answer for current question and asparagus piece
+        Args:
+            answer: The answer to be logged for the current question. It will be writtent to the outputfile once you jump to the next piece.
+
+        """
+        self.set_value_for_label(answer,self.questions[self.idx_question])
+
+    def write_answers_to_file(self):
+        """ Writes answer to questions for current idx to file """
+        try:
+            df = pd.DataFrame.from_dict(self.labels, orient = "index", columns=self.outfile_headers)
+            df.to_csv(self.outpath,sep =";")
+        except:
+            QMessageBox.about(self, "Attention", "Writing file failed.")
+
+    def previous_question(self):
+        """ Changes index to previous file and draws respective asparagus"""
+        if(self.idx_question==0):
+            self.previous_image()
+            self.idx_question = len(self.questions)-1
+        else:
+            self.idx_question -= 1
+        self.ui.question.setText(self.questions[self.idx_question])
+
+    def next_question(self):
+        """ Changes index to next file and draws respective asparagus"""
+        if(self.idx_question==len(self.questions)-1):
+            if self.thread.isRunning():
+                QMessageBox.about(self, "Attention", "Feature extraction still running")
+                return
+            self.next_image()
+            self.idx_question = 0
+        else:
+            self.idx_question += 1
+        self.ui.question.setText(self.questions[self.idx_question])
+
+    def rek_get_files(self, path, regex):
+        """ Gets files that match regex for the specified path and all subdirectories recursively.
+        Args:
+            path: A path to the parent directory.S
+            regex: A regex that specifies the name of files searched for.
+
+        """
+        for f in os.scandir(path):
+            if f.is_dir():
+                self.rek_get_files(f.path, regex)
+            else:
+                if re.match(regex,f.name):
+                    self.files.append(path+"/"+f.name)
+
+    def set_filenames(self, directory):
+        """ Sets attribute self.files and draws asparagus piece
+        Args:
+            directory: File directory that contains images satisfying the naming convention in any arbitrarily nested subdirectories.
+        """
+
+        self.files = []#All files in subtree
+        self.rek_get_files(directory,".*\.png")#Traverse subdirectories & get all .bmp filepaths
+
+        ids_to_files = {}
+
+        for path in self.files:
+            match = re.search(".*/([0-9]+)_[a-z]\.png",path)
+            if match:#Get id using regex.
+                id = int(match.groups()[0])
+                try:#Create list as key of id_to_files if it doesn't exist already.
+                    ids_to_files[id]
+                except:
+                    ids_to_files[id] = []
+                ids_to_files[id].append(path)#Append filename to list.
+
+        self.images = ids_to_files#list(ids_to_files.values())
+        self.idx_image = 0
+        self.draw_asparagus()
+
+    def next_image(self):
+        """ Increases self.idx_image if this index exists. Important: This method is the only one that writes the answet to the file."""
+        self.write_answers_to_file()
+        self.set_index(self.idx_image+1)
+        if not self.idx_image in self.images:#In case of failure a message is printed elsewhere
+            self.set_index(self.idx_image-1)
+
+    def previous_image(self):
+        """ Decreases self.idx_image if this index exists"""
+        self.set_index(self.idx_image-1)
+        if not self.idx_image in self.images:#In case of failure a message is printed elsewhere
+            self.set_index(self.idx_image+1)
+
+    class Features(QThread):
+        color_plot = pyqtSignal(np.ndarray)
+        predictionWidth = pyqtSignal(str)
+        predictionBended = pyqtSignal(str)
+        predictionLength = pyqtSignal(str)
+        predictionViolet = pyqtSignal(str)
+        predictionRust = pyqtSignal(str)
+        predictionBlooming = pyqtSignal(str)
+
+        predictionWidth_2 = pyqtSignal(str)
+        predictionBended_2 = pyqtSignal(str)
+        predictionLength_2 = pyqtSignal(str)
+        predictionViolet_2 = pyqtSignal(str)
+        predictionRust_2 = pyqtSignal(str)
+        predictionBlooming_2 = pyqtSignal(str)
+
+        predictionWidth_3 = pyqtSignal(str)
+        predictionBended_3 = pyqtSignal(str)
+        predictionLength_3 = pyqtSignal(str)
+        predictionViolet_3 = pyqtSignal(str)
+        predictionRust_3 = pyqtSignal(str)
+        predictionBlooming_3 = pyqtSignal(str)
+
+        overallPredictionWidth = pyqtSignal(str)
+        overallPredictionBended = pyqtSignal(str)
+        overallPredictionLength = pyqtSignal(str)
+
+        overallPredictionViolet = pyqtSignal(str)
+        overallPredictedValueRust = pyqtSignal(str)
+        overallPredictionBlooming = pyqtSignal(str)
+
+        def __init__(self, outer):
+            """ Initializes the feature extracting thread
+            Note that setting features from the thread directly is possible in general as the "factory method" is used to init the thread (self/self.ui is passed to this).
+            However setting values to the ui is not inteded by PyQt. To eliminate all potential problems due to parallel updating of values slots/signals are used for communication
+            between the thread and self.outer (signals are handled in a queued connection).
+
+            As self.idx_image is used for updating values in self one shall not be able to jump to the next_image (otherwise saving for incorrect image index).
+            Hence respective functions are deactivated as long as the thread is running in the methods of LabelingDialog.
+            Args:
+                outer: Outer instance of class LabelingDialog
+            """
+            super().__init__()
+            self.outer = outer
+
+        def run(self):
+            """ Run method of the QThread that is started in parallel when QThread.start() is called. Extracts features emits results as signals. """
+            try:
+                imgs = [np.array(imageio.imread(fname)) for fname in self.outer.images[self.outer.idx_image]]
+                idx_image = self.outer.idx_image#At creation time
+            except KeyError:
+                #If there are no images for the index for whatever reason. Just dont do anything.
+                self.outer.ui.asparagus_number.setEnabled(True)
+                return
+
+            try:
+                self.color_plot.emit(color_plot(imgs))
+            except:
+                self.outer.ui.asparagus_number.setEnabled(True)
+                return
+
+            try:
+                p = [estimate_width(np.rot90(x)) for x in imgs]
+                self.predictionWidth.emit(str(int(p[0])))#Numerical widthprint('% 6.2f' % v)
+                self.predictionWidth_2.emit(str(int(p[1])))
+                self.predictionWidth_3.emit(str(int(p[2])))
+                width = np.round(np.mean(np.array(p)))
+                self.overallPredictionWidth.emit(str(width))
+            except Exception as e:
+                print(traceback.format_exc())
+                print("Couldn't set auto width")
+
+            try:
+                p = [estimate_purple(x, threshold_purple=10) for x in imgs]
+                self.predictionViolet.emit(str(p[0]))#Numerical widthprint('% 6.2f' % v)
+                self.predictionViolet_2.emit(str(p[1]))
+                self.predictionViolet_3.emit(str(p[2]))
+                #most_common = Counter(np.array(p)).most_common(1)[0][0]
+                purple_estimate = np.round(np.mean(np.array(p)))
+                self.overallPredictionViolet.emit(str(purple_estimate))
+            except Exception as e:
+                print(traceback.format_exc())
+                print("Couldn't set auto purple")
+
+            try:
+                p = [estimate_bended(x) for x in imgs]
+                self.predictionBended.emit(str(int(p[0])))#'{:10.1}'.format(p[0][1]))
+                self.predictionBended_2.emit(str(int(p[1])))#'{:10.1}'.format(p[1][1]))
+                self.predictionBended_3.emit(str(int(p[2])))#'{:10.1}'.format(p[2][1]))
+                #is_bended = np.sum(np.array(p)[:,0])>1#If at least one image shows it's bended
+                bended = np.round(np.mean(np.array(p)))
+                self.overallPredictionBended.emit(str(bended))
+            except Exception as e:
+                print(traceback.format_exc())
+                print("Couldn't set auto bended")
+
+            try:
+                p = [estimate_length(x) for x in imgs]
+                self.predictionLength.emit(str(int(p[0])))
+                self.predictionLength_2.emit(str(int(p[1])))
+                self.predictionLength_3.emit(str(int(p[2])))
+                length = np.round(np.mean(np.array(p)))
+                self.overallPredictionLength.emit(str(length))
+            except Exception as e:
+                print(traceback.format_exc())
+                print("Couldn't set auto length")
+
+            self.outer.ui.asparagus_number.setEnabled(True)
+
+    def draw_asparagus(self):
+        """ Draws image of asparagus pieces from three perspectives"""
+        try:
+            imgs = []
+            max_y = 0
+            max_x = 0
+            for i, fname in enumerate(self.images[self.idx_image]):
+                im = imageio.imread(fname)
+                im = np.rot90(im).copy()
+                imgs.append(im)
+
+                max_y += im.shape[0]
+                if im.shape[1] > max_x:
+                    max_x = im.shape[1]
+
+            n_channels = imgs[0].shape[2]
+            combined = np.zeros([max_y,max_x,n_channels],dtype=np.uint8)
+
+            y_offset = 0
+            for im in imgs:
+                combined[y_offset:y_offset+im.shape[0],:im.shape[1],:] = im
+                y_offset += im.shape[0]
+
+            self.ui.label.update(np.rot90(combined,3).copy())
+        except KeyError:
+            QMessageBox.about(self, "Attention", "No images found for current index (" + str(self.idx_image) + ")")
+        except Exception as e:
+            print(e)
+            return
+
+    def update_info(self):
+        try:
+            msg = ""
+            for img_path in self.images[self.idx_image]: # change to idx??
+                img_name = re.search(".*/(.*_[a-z]\.png)",img_path).groups()[0]
+                msg += " | " + img_name + " | "
+            self.ui.asparagus_name.setText(msg)
+        except:
+            pass
+
+    def set_index(self,idx):
+        """ This method sets the current index and starts feature extraction if self.extract_features is True
+            (A message is displayed via self.draw_asparagus if no respective image exists).
+            Important: This method does not save results of feature extraction of answered questions to the output csv file.
+            This allows the user to jump to arbitrary indices without inserting empty lines to the csv.
+        Args:
+            idx: The index
+        """
+        if(self.thread.isRunning()):
+            QMessageBox.about(self, "Attention", "Feature extraction still running")
+            return
+        self.idx_image = idx
+        self.draw_asparagus()
+        self.update_info()
+
+        #The following lines are necessary as otherwise another signal would be emitted that would further increase the index
+        self.ui.asparagus_number.blockSignals(True)
+        self.ui.asparagus_number.setValue(self.idx_image)
+        self.ui.asparagus_number.blockSignals(False)
+
+        self.idx_question = 0
+        self.ui.question.setText(self.questions[self.idx_question])
+
+        if self.extract_features:
+            self.ui.asparagus_number.setEnabled(False)
+            self.thread.start()
+
+    def eventFilter(self, source, event):
+        """ Filters key events such that arrow keys may be handled.
+            Args:
+                source: Source of event
+                event: Event to be handled
+        """
+        if event.type() == QtCore.QEvent.KeyRelease:
+            id_right = 16777236
+            id_left = 16777234
+
+            if event.key() == id_right:
+                self.answer_no()
+
+            elif event.key() == id_left:
+                self.answer_yes()
+
+        try:#When closing the app the widget handled might already have been destroyed
+            return self.widget_handled.eventFilter(source, event)#Execute the default actions for the event
+        except:
+            return True#a true value prevents the event from being sent on to other objects
